+++ GodotEngine.Utility/EditorSerialization/BuildSerialization.cs
@@ -3,7 +3,7 @@ using System;
 using System.Reflection;
 using System.Collections.Generic;
 
-namespace Cobilas.GodotEngine.Utility.Serialization;
+namespace Cobilas.GodotEngine.Utility.EditorSerialization;
 /// <summary>Class allows to build a serialization list of properties of a node class.</summary>
 public static class BuildSerialization {
     private const BindingFlags flags = BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic;