+++ GodotEngine.Utility/EditorSerialization/PropertyCustomAttribute.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Cobilas.GodotEngine.Utility.Serialization;
+namespace Cobilas.GodotEngine.Utility.EditorSerialization;
 /// <summary>This attribute indicates to <seealso cref="BuildSerialization"/> 
 /// that the <seealso cref="PropertyCustom"/> class belongs to a certain type.</summary>
 [AttributeUsage(AttributeTargets.Class, Inherited = false, AllowMultiple = true)]