+++ GodotEngine.Utility/EditorSerialization/ShowPropertyAttribute.cs
@@ -1,18 +1,19 @@
 using Godot;
 using System;
 
-namespace Cobilas.GodotEngine.Utility.Serialization; 
-/// <summary>The attribute allows you to hide and save the value of a field or property in the editor.</summary>
+namespace Cobilas.GodotEngine.Utility.EditorSerialization;
+/// <summary>The attribute allows you to show a field or property in the editor.</summary>
 [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property, AllowMultiple = false, Inherited = false)]
-public class HidePropertyAttribute : SerializeFieldAttribute {
+public class ShowPropertyAttribute : SerializeFieldAttribute {
     /// <inheritdoc/>
     public override bool SaveInCache { get; protected set; }
     /// <inheritdoc/>
     public override PropertyUsageFlags Flags { get; protected set; }
     /// <inheritdoc cref="SerializeFieldAttribute(PropertyUsageFlags, bool)"/>
-    public HidePropertyAttribute(bool saveInCache) : base(
+    public ShowPropertyAttribute(bool saveInCache) : base(
         PropertyUsageFlags.ScriptVariable |
-        PropertyUsageFlags.Storage, saveInCache) {}
+        PropertyUsageFlags.Storage |
+        PropertyUsageFlags.Editor, saveInCache) {}
     /// <inheritdoc cref="SerializeFieldAttribute(PropertyUsageFlags, bool)"/>
-    public HidePropertyAttribute() : this(false) {}
+    public ShowPropertyAttribute() : this(false) {}
 }